Renamed telnyxSend to send and move to sms folder
